+++ manual/source/version_history.rst
@@ -13,6 +13,31 @@ if you'd like to give us feedback or help in the development.  See :ref:`support
 * Multiple beam line tracking.
 * Use sampler data from a BDSIM output file as input to another BDSIM simulation.
 
+
+V1.6.0 - 2021 / XX / XX
+=======================
+
+New Features
+------------
+
+* New skimming tool called :code:`bdskim` is included for skimming raw data. See :ref:`bdskim-tool`.
+* New combination tool called :code:`bdsimCombine` is included to merge raw data files
+  and skimmed data files alike. See :ref:`bdsimCombine-tool`.
+
+General
+-------
+
+Bug Fixes
+---------
+
+Output Changes
+--------------
+
+
+Output Class Versions
+---------------------
+
+
 V1.5.1 - 2020 / 12 / 21
 =======================